Use doubles instead of ints to prevent integer overflow

+++ src/variables-extern.c
@@ -76,8 +76,8 @@ extern gchar *dernier_chemin_de_travail;
 extern gint nb_derniers_fichiers_ouverts;          /* contient le nb de derniers fichiers ouverts */
 extern gint nb_max_derniers_fichiers_ouverts;        /* contient le nb max que peut contenir nb_derniers_fichiers_ouverts ( réglé dans les paramètres ) */
 extern gchar **tab_noms_derniers_fichiers_ouverts;
-extern gint compression_fichier;
-extern gint compression_backup;
+extern gdouble compression_fichier;
+extern gdouble compression_backup;
